Initial commit or rust-rm
This is the first commit of the _rust-rm_ project, a Rust implementation
of an `rm(1)` [1] Unix command-like CLI. The goal of this `rm` command
is to feel familiar yet be more safe and user friendly. Following [2, 3]
it improves upon `rm(1)` with a (default) dry run mode, suggestions for
next steps in its output, and a more modern CLI experience (thanks to
the `clap` crate [4]). It also adds support for moving to the trash
instead (thanks to the `trash` crate [5]).

The project is formatted using rustfmt, extensively tested, documented,
and analyzed. Effort has been put in to make the design extendable and
open for changes, one way in which this manifests is the effective way
in which compile-time features are integrated into the implementation.
Similarly, effort has been put in to write tests and make writing tests
simple.

All of this is continuously verified using a Continuous Integration (CI)
setup powered by GitHub Actions. Note that commands used in CI are
available verbatim during local development to allow contributors to
conveniently verify their changes will (likely) pass CI checks.

This project is initialized for open source use and contributions with a
license (MPL-2.0, chosen as this is end-user software and not a library)
and docs for both users (README.md) and contributors (CONTRIBUTING.md).
A security policy is also included to welcome security researches to
look at the project.

--
1. https://man7.org/linux/man-pages/man1/rm.1.html
2. https://clig.dev/
3. https://git-scm.com/docs/git-clean
4. https://crates.io/crates/clap
5. https://crates.io/crates/trash