Remove the `no_color.rs` integration test suite
Update the integration tests to exclude the suite for color/ansi
formatting testing. The test suite currently doesn't function - removing
line 51 yields the same test results, proving the `NO_COLOR` environment
variable is not what's suppressing ANSI formatting in the suite.

The test is being removed here as a precaution to avoid false confidence
that the suite provides any guarantees w.r.t. `NO_COLOR` support by this
application. As of this commit no alternative testing strategy for this
purpose was found.