Add ANSI styling to program output
Extend the use of ANSI styling from help messaging (generated by `clap`
[1]) to the program's own output. ANSI styling is implemented using
`anstream` [2] (also used by `clap`) and `owo-colors` [3] (suggested in
the documentation of `anstream`).

Some ANSI styling is applied to most output, mainly with the goal of
highlighting the most important information and dimming optional output.
In particular:
- Highlight what file/directory a message is about.
- Dim all tips.
- Dim all verbose messages.

Additionally, this somewhat correctly omits the `CLEARLINE` in
`--interactive` mode when ANSI isn't supported. Unfortunately, it does
not necessarily make sense that the `CLEARLINE` is now also omitted
with, e.g., `NO_COLOR=1` because it's not _actually_ color.

--
1. https://crates.io/crates/clap
2. https://crates.io/crates/anstream
3. https://crates.io/crates/owo-colors