Update the `--help` sentence about removing directories
Correct the sentence in the `--help` text about removing directories. It
now, correctly, says `--dir` will remove empty directories and
`--recursive` will remove non-empty directories.

The previous text was based on an experimental implementation prior to
v23.04 where `--recursive` would only clear the contents of directories,
which had to be combined with `--dir` to removing a directory including
its content.