Remove filled directories in one go with `--trash --recursive`
Update the behavior of the software when both the `--trash` and the
`--recursive` options are used to delete filled directories as a whole.
This is instead of recursively deleting the contents of filled
directories, which is still the behavior without the `--trash` option.

This change is implemented because of two reasons. First, moving to
trash is quite a different than outright removing files, thus removing a
filled directory isn't as concerning. Second, restoring recursively
removed directory contents can be inconvenient depending on how the
trash bin operates, potentially being impossible unless the containing
directory is restored first.

On the positive side, this change allows for the removal of the TOCTOU
empty directory check in `rm::dispose`. This check also in a way
motivates the change to begin with - the `trash` crate [1] takes the
stance that moving filled directories in one go isn't a problem.

A perhaps surprising change (on the negative side...) is hte mapping of
opened `fs::Entry`s to their visited version in `walk::given`. This
change is necessary for the scenario tested in the new test in the
`--interactive` test suite: avoid asking to descend into a filled
directory if we're not going to. I'm not a huge fan of this approach,
but to be honest I'm not a big fan of the `visited` field on `fs::Entry`
to begin with.

--
1. https://crates.io/crates/trash