Proof of concept deletion of `is_empty(&entry)` check when removing dirs
Enable the `io_error_more` feature gate to get access to enum variant
`io::ErrorKind::DirectoryNotEmpty` so that trying to remove a filled
directory (due to a TOCTOU mismatch) fails correctly without the custom
emptiness check in the `rm::remove` function.

The `channel` in the rust-toolchain.toml file was updated to make sure
all invoked cargo commands use the nightly toolchain. THIS IS A
TEMPORARY CHANGE INTENDED ONLY TO HELP DEMONSTRATE THE PROOF OF CONCEPT!