Remove minimum supported rust version (MSRV) check
Update the continuous integration (CI) based check for the minimum
supported rust version (MSRV) [1] as well as the corresponding optional
prerequisite from the Contributing Guidelines.

While an interesting check, the MSRV of this project does not appear to
be relevant to me as this is not a library. In general, especially with
with the tooling provided by the Rust ecosystem, contributors and users
of this software can easily access and use the latest stable version of
Rust to compile (or more, e.g. test) it. Combined, these reasons justify
not checking for an MSRV.

Correspondingly, remove the (minimum) `rust-version` field from the
manifest [2] and update the Rust/Cargo version noted in the README.md
to match that of the `rust-toolchain.toml` file.

--
1. https://github.com/foresterre/cargo-msrv
2. https://doc.rust-lang.org/cargo/reference/manifest.html