Correct stated license, add documentation license
Update the README.md with a correction to the state license. It
incorrectly stated this project is licensed under the MPL-2.0 license
(this was a copy-paste error from one of my other projects) while
linking to this project's Apache-2.0 license text. The README.md
statement and LICENSE file (as well as source code SPDX license ids)
are now aligned on Apache-2.0

Also add an explicit license for documentation content. A "permissive"
creative commens variant (as opposed to the "copyleft" Share-Alike
variant) is used in line with the permissive source code license.