Updates for Clippy 1.70.0
Address errors from Clippy that started with 1.70.0 and also update the
Clippy configuration given changes from 1.69.0 to 1.70.0.

The source code changes are due to the new `manual_main_separator_str`
lint (category: `complexity`). The rule `let_underscore_untyped` has
been explicitly enabled because it was previously enabled because it was
in the `pedantic` category but has now been moved to `restriction`. None
of the other new rules were explicitly enabled because they're either:
1. in a category that is generally enabled,
2. in the `nursery` category and don't apply to this project, or
3. in the `restriction` category and don't apply to this project.