Set up mutation testing with cargo-mutants
Initialize the mutation testing for this project using cargo-mutants
[1,2] as mutation testing framework. The goal of using mutation testing
is to improve the unit tests of this project and keep them good over
time by making sure changes to the source code are properly tested.

First, added a new Just command to run mutation tests with the correct
configuration (and a corresponding `ci-` variant). This configures the
runner to:
1. Output reports into the `_reports/mutants.out/` directory, similar to
   how coverage is put in the `_reports/` folder. Correspondingly update
   coverage commands to output into `_reports/coverage/` folder for
   clearer separation.
2. Exclude certain functionality:
   - The `cli::run` function because it isn't unit tested.
   - The `logging` module because it isn't unit tested.
3. Run only the unit tests.
4. Use the configured `test_features`.

Second, update the Contributing Guidelines to include cargo-mutants as
an optional prerequisite and include a section on mutation testing. Like
test coverage, the mutation testing section describes why we do it and
how to run it.

Third, update the continuous integration Rust workflow to include a job
for running mutation tests and upload the report for easy reference. The
job serves the same purpose as the coverage job does but for mutation
testing instead, hence the similarities.

The choice of mutation testing using only unit tests is primarily to
ensure mutants are caught early and cheaply when they occur, whether by
mutation testing or by changing the code. The integration tests for this
project should be concerned with ensuring the CLI as a whole works as
expected, not whether mutations are caught. That's not to say mutation
testing can't help improve integration tests (e.g. through mutants in
ignored functions), just that it's not the primary goal.

As of this commit there's two surviving mutants:
1. replace `fs::<impl Display for ErrorKind>::fmt -> std::fmt::Result`

   Because the exact strings aren't asserted in the unit tests (only in
   the integration tests). How to address this is TBD.

2. replace `rm::dispose -> Result`

   Because running mutation tests with `test-trash` enabled currently
   fails on the `Entry::path` mutant (which changes it's implementation
   to `Default::default()`). With that mutant some of the `test-trash`
   tests will delete the current working directory, causing a panic. How
   to address this is TBD.

--
1. https://github.com/sourcefrog/cargo-mutants
2. https://mutants.rs/welcome.html