Guard against disposing the CWD & exclude `impl Display` mutants
This commit is based on the mutation testing results. First, exclude
mutants on `std::fmt::Display` implementation as we don't want to
(necessarily) test the content of those strings in unit tests (if there
are ideas for how to effectively and meaningfully test these, this
exclusion can be removed). Second, run mutation tests with `test-trash`
enabled so that mutants in trash-related code can be caught. Third, and
this follows from problems observed after enabling `test-trash` for
mutation testing (details below), guard against deleting the current
working directory when disposing of entries to the trash. In particular,
check if the path is empty (`""`) and refuse to remove if it is.

This guard was added following a the mutant on `Entry::path`:

      pub fn path(&self) -> PathBuf {
    -    Path::new(&self.path).to_owned()
    +    Default::default()
      }

which caused a panic because the `Default::default()` is an empty path
(`""`) and the trash crate's behavior is such that it removes the
current working directory given an empty path. This behaviour is not
desired, hence the extra guard.

The guard is "extra" because in general this software should already
catch and error given the empty path `""`. But without a explicit check
I hesitate to leave this potentially dangerous behavior open. Note that
otherwise deleting the current working directory is explicitly refused
as well.

I tested what happens when calling `std::fs::{remove_dir,remove_file}`
with `""` and verified that those will error with a `NotFound` error.