Set up continuous coverage with Codecov
Adjust the "Rust / Coverage" job to continuously upload coverage data to
Codecov [1] in order to gain insight over coverage over time.

To achieve this, first a `just ci-coverage` command is created that runs
coverage with the correct configuration for CI usage (like other `ci-*`
commands this is available for local execution as well). This changes
`just coverage` to:
1. Run with all test flags enabled for maximum coverage, and
2. Output an lcov file (in addition to HTML) to upload to Codecov.

The "Rust / Coverage" job is updated to use the new `just ci-coverage`
command as well as to "Upload coverage to codecov" using the new lcov
coverage file. (For now, the HTML report artifact upload is kept for
more universal access to coverage results.)

Lastly, a Codecov configuration file is added to set requirements and
integration behavior.

--
1. https://codecov.io