Ignore potential items related to used tooling
Update the `.gitignore` to include ignore directives for standard items
that could be created by tooling if invoked without the `Justfile`. This
is meant as a preventative measure to avoid these accidentally being
commit when a contributor is testing with used tooling but, say, new
configurations. Correspondingly update the `just clean` command.

The chosen patterns come from the following tools (tool list taken from
the Contributing Guidelines):
- `cargo-all-features`: None
- `cargo-deny`: None
- `cargo-mutants`: "/mutants.out*/", for mutants.out and mutants.out.old
- `cargo-tarpaulin`: "cobertura*", "lcov*", and "tarpaulin*" for the
  reports outputted by each supported `--out` format (XML, LCOV, HTML,
  and JSON)