Improve continuous integration artifacts
Update the continuous integration (CI) workflows to improve the use of
and maintain artifacts.

First, make sure the artifacts are uploaded in case of a job failure.
A job failure due to low coverage or uncaught mutants are some of the
most important reports to be available as they should be addressed - and
not having to rerun the command locally is very useful.

Second, explicitly configure the retention time for artifacts. This is
mostly just to avoid unexpected changes due to a dependency update as
well as to set clear expectations for how long the artifact will be
available.