Rename classic mode to GNU mode
Change the name of the mode that emulates the ~~Unix~~ GNU version of
rm(1) from classic mode (RUST_RM_CLASSIC) to GNU mode (RUST_RM_GNU_MODE)
in order to 1) make it easier to guess the behavior when the environment
variable is set, and 2) make it explicit what behavior will be emulated.

Note that, prior to this change, per the first change in the README.md,
it was understood that classic mode emulated the "Unix" implementation
of rm(1). This is not the case (and I don't think a  "Unix" version of
rm(1) actually exists), instead it emulates the GNU version of rm [1].
As such, the new name was chosen per the above reasoning.

A side benefit of using a name that explicitly references what is being
emulated is that it leaves open the possibility of adding support for
another emulation mode in the future (e.g. POSIX mode).

All references to "classic mode" and "Unix rm(1)" should have been
updated in this commit. If any such reference lingers feel free to
remove it at your convenience.

--
1. https://man7.org/linux/man-pages/man1/rm.1.html