Make `CurrentDirPath` and `ParentDirPath` more user friendly
Update the test utilities `CurrentDirPath` and `ParentDirPath` to be
more use friendly by avoiding the need to modify them with the
`MAIN_SEPARATOR(_STR)` at time of use. Instead, it will now generate
correct paths for the platform where tests are run by leveraging the
`MAIN_SEPARATOR_STR` in the arbitrary.