Ignore all tests with symlinks for Windows by default
Add conditional ignore attributes to all tests involving symlinks to
prevent them from running on Windows. Per [1,2], creating symlinks is
considered a privileged action on Windows systems. As a result, all
these tests currently fail on Windows systems unless they're run as
administrator (or similar)[^1].

To address this, the test_feature "test-symlink" has been created to
allow Windows contributors to run tests involving symbolic links
conditionally. This is documented in the Contributing Guidelines. This
test_feature is enabled by default when running `ci-*` variants of Just
commands (like other test_features).

Note that tests with pre-existing ignores for Windows due to symlinks
(i.e. those with the ignore comment "TODO: investigate symlink test
errors on Windows") are not included here. This is because those tests
fail due to a bug in the implementation of this software for Windows
where symbolic links to folders cannot be removed.

--
^1: This was not the case in the CI because (I presume) GitHub Actions
    runs things as administrator (or similar) by default.
--
1. https://doc.rust-lang.org/stable/std/os/windows/fs/fn.symlink_dir.html
2. https://doc.rust-lang.org/stable/std/os/windows/fs/fn.symlink_file.html