Bump Rust toolchain channel from 1.70 to 1.71
Upgrade the version of Rust and related tooling from 1.70.0 to 1.71.0,
which was recently released[1].

The `io_error_more` gate is still applicable (verified based on the
corresponding GitHub issue and running the tests with the related code
removed).

The Clippy configuration has been updated to explicity enable a new
rule from the "restriction" category (in addition to other rules that
are enabled by virtue of enabling other categories). No violations are
found by Clippy as a result of this toolchain bump.

--
1. https://blog.rust-lang.org/2023/07/13/Rust-1.71.0.html