Improve `test_disallow_root` suite for Windows
First, adjust the `prop_assume!` for the `test_disallow_root::non_root`
to correctly filter out any path that is a root path on all systems,
including Windows. This fixes a flakiness problem with this test on
Windows.

Secondly, add a Windows-specific test to that a path consisting of only
the drive is considered a file system root. To achieve this, the suite
has been updated to use a parameterized test approach (as described in
the Contributing Guidelines).