Set up GitHub Actions workflow to automatically create a GitHub Release
Create a new GitHub Actions workflow to automatically create a GitHub
Release when a (release) tag is pushed to the GitHub repository. This
workflow, inspired by [1], uses cargo and cross [2] to compile the
project for five targets (3 platforms, 2 architectures) in one set of
jobs and create a GitHub Release with those compiled artifacts in a
second (dependent) job.

--
1. https://dzfrias.dev/blog/deploy-rust-cross-platform-github-actions
2. https://github.com/cross-rs/cross