Bump `cargo-mutants` and `cargo-tarpaulin` to latest
Update all cargo tooling (listed in the prerequisites section of the
Contributing Guidelines) to the latest version. `cargo-all-features` and
`cargo-deny` don't have a new version available. `cargo-mutants` bumped
from 23.5.0 to 23.6.0 with a myriad of improvements. `cargo-tarpualin`
bumped from 0.25.2 to 0.26.1 with some fixes and improvements.

Since no new features are utilized the minimum versions of these tools
listed in the Contributing Guidelines have not been updated.

Besides staying up-to-date, these updates should help improve the
visibility into the quality of the tests of this project.