Bump Rust toolchain channel from 1.72 to 1.73
Upgrade the version of Rust and related tooling from 1.72.0 to 1.73.0,
which was recently released[1].

The `io_error_more` gate is still applicable (verified based on the
corresponding GitHub issue and running the tests with the related code
removed).

Also updated the Clippy configuration - both enabled lints in Justfile
and individual lint configurations in clippy.toml - in accordance with
the changes of this new version. Note that the `ignored_unit_patterns`
lint has been disabled explicitly because it reports "incorrectly" for
various uses of macros.

All changes in code are due to Clippy lints:
- Changes to `use` statements and fully qualified expressions are due to
  the `absolute_paths` lint. This lint was mostly useful, but has been
  configured to ignore the `log` and `trash` crates since those aren't
  as readable without the fully qualified import path. In some cases
  this change also just makes the codebase more consistent.
- Changes from `Err(_)` to `Err(())` where the error type is Unit are
  due to the `ignored_unit_patterns` lint. Even though it was disabled
  due to noise, it did suggest these useful changes. To elaborate, I
  believe this change is useful because, per the lint description [2]:

  > [...] it would detect a type change that `_` would ignore.

--
1. https://blog.rust-lang.org/2023/10/05/Rust-1.73.0.html
2. https://rust-lang.github.io/rust-clippy/master/index.html#/ignored_unit_patterns