Bump Rust toolchain channel from 1.73 to 1.74
Upgrade the version of Rust and related tooling from 1.73.0 to 1.74.0,
which was recently released[1].

The `io_error_more` gate is still applicable (verified based on the
corresponding GitHub issue and running the tests with the related code
removed).

No updates were necessary related to rust-fmt or Clippy this time. The
formatting is still adhered to and there are no new not-enabled Clippy
rules that aren't deprecated or in the nursery.

--
1. https://blog.rust-lang.org/2023/11/16/Rust-1.74.0.html