Bump Rust toolchain channel from 1.74 to 1.75
Upgrade the version of Rust and related tooling from 1.74.0 to 1.75.0,
which was recently released[1].

The `io_error_more` gate is still applicable (verified based on the
corresponding GitHub issue and running the tests with the related code
removed).

No updates were necessary related to rust-fmt or Clippy this time. The
formatting is still adhered to and there are no new not-enabled Clippy
rules that aren't deprecated or in the nursery.

Cargo now allows omitting the `package.version` field in Cargo.toml[2],
so since this project isn't published (and the version isn't tracked
using that field), it is removed.

The mutation testing configuration has been updated to ignore the main
function, which gained a viable mutant due to this upgrade, but isn't
covered by the unit tests (which is the scope of mutation testing).

--
1. https://blog.rust-lang.org/2023/12/28/Rust-1.75.0.html
2. https://github.com/rust-lang/cargo/blob/master/CHANGELOG.md