Bump Rust toolchain channel from 1.75 to 1.76
Upgrade the version of Rust and related tooling from 1.75.0 to 1.76.0,
which was recently released[1].

The `io_error_more` gate is still applicable (verified based on the
corresponding GitHub issue and running the tests with the related code
removed).

No updates were necessary related to rust-fmt. The Clippy configuration
has been updated to enable some of the new lints in the "restriction"
category that make sense to me and can be enabled easily for this
project.

--
1. https://blog.rust-lang.org/2024/02/08/Rust-1.76.0.html