Bump Rust toolchain channel from 1.76 to 1.77
Upgrade the version of Rust and related tooling from 1.76.0 to 1.77.0,
which was recently released[1].

The `io_error_more` gate is still applicable (verified based on the
corresponding GitHub issue and running the tests with the related code
removed).

No updates were necessary related to rust-fmt. The Clippy configuration
has been updated to enable some of the new lints in the "restriction"
category that make sense and can be enabled easily for this project.
Also, the `allowed-duplicate-crates` config option was added to address
some new warnings from the `clippy::multiple_crate_versions` rule for
development dependencies.

--
1. https://blog.rust-lang.org/2024/03/21/Rust-1.77.0.html