Define a development image
Add a `Containerfile` for a development image that will come installed
with all necessary and optional project dependencies for easier
development.

The `Containerfile` is intentionally licensed under the MIT-0 (MIT No
Attribution) license (more permissive than this project's default
Apache-2.0 license) so as to make using it as a base for a development
image of another project easier - no attribution required.

This comes with upgrades to two tooling updates because the previously
used version where (for some reason) incompatible with the container.

As a basic continuous check, a CI job has been added to build the
container image to ensure at least that works. Unfortunately this leaves
the possibility that some command won't work inside the ephemeral
development environment.