Bump Rust toolchain channel from 1.79 to 1.80
Upgrade the version of Rust and related tooling from 1.79.0 to 1.80.0,
which was recently released [1].

The `io_error_more` gate is still applicable (verified based on the
corresponding GitHub issue and running the tests with the related code
removed).

One new clippy lints from the "restriction" category ise enabled for
this project because I think it's a sensible restriction. Also, the new
`unexpected_cfgs` lint has been explicitly configured so that it
1) errors instead of warns and 2) does not trigger for the tarpaulin
config used by this project.

--
1. https://blog.rust-lang.org/2024/07/25/Rust-1.80.0.html