Bump Rust toolchain channel from 1.80 to 1.81
Upgrade the version of Rust and related tooling from 1.80.0 to 1.81.0,
which was recently released [1]. Additionally, upgrade the version of
Alpine used for the development container from 1.19 to 1.20

The `io_error_more` gate is still applicable - verified based on the
corresponding GitHub issue and running the tests with the related code
removed.

This update of clippy introduced a requirement to inline the reason for
exceptions, hence this was added here. Interestingly, this lead to two
refactorings to avoid the exception leading to cleaner code (in test
`walk::test_visit::transform` and `walk::test_visit::skip`) as well as
one unnecessary exception. Additionally, two new lints from the
restriction category were explicitly enabled as they are considered
applicable to this project.

--
1. https://blog.rust-lang.org/2024/09/05/Rust-1.81.0.html