Bump Rust toolchain channel from 1.81 to 1.82
Upgrade the version of Rust and related tooling from 1.81.0 to 1.82.0,
which was recently released [1]. Additionally, this sets the version in
the project manifest (Cargo.toml) for completeness (found via the new
`cargo info` command).

The `io_error_more` gate is still applicable - verified based on the
corresponding GitHub issue and running the tests with the related code
removed. However, it seems 1.83 will be the end of this, see [2].

For the update to Clippy, two new rules from the "Restrictions" group
were explicitly enabled as they're relevant to this project. I also re-
evaluated the allow directive for the `ignore_unit_patterns` lint and
decided to just refactor the codebase in line with the lint instead
(it was allowed in 44b36a501616bcd8ae4355095181215e424c13d3 originally,
based on the reason given there I *guess* the noisiness has been reduced
by Clippy in the meanwhile).

Lastly, this fixes a problem in the ephemeral development environment
where some `just` commands would fail because git did not recognize the
working directory as trusted. This has been addressed by telling git to
trust it.

--
1. https://blog.rust-lang.org/2024/10/17/Rust-1.82.0.html
2. rust-lang/rust#131396