Bump cargo-tarpaulin from v0.28.0 to v0.31.2
Upgrade the version of cargo-tarpaulin used for coverage purposes to the
latest available release. This is in response to failures when using
cargo-tarpaulin v0.28.0 starting with Rust 1.82.

The latest version of cargo-tarpaulin, v0.31.2, is now set as the
version used as well as the version recommended on the Contributing
Guidelines. This is in order to ensure everyone (following the project
guidelines) is able to run all commands. It _could_ be that an earlier
version of cargo-tarpaulin is also compatible with Rust 1.82, but I did
not bother to find the earliest version (if you know, don't hesitate to
let me know and we can update the Contributing Guidelines accordingly).