Bump Rust toolchain channel from 1.82 to 1.83
Upgrade the version of Rust and related tooling from 1.82.0 to 1.83.0,
which was recently released [1]. Additionally, this updates the version
of the lockfile from 3 to 4 (along with some dependency updates).

This release or Rust removed the `io_error_more` gate from the directory
not empty error kind, hence the manual check is removed and conversion
logic has been added to the `From<io::ErrorKind>` trait implementation
of `fs:ErrorKind`.

For Clippy, on new optional rule from the "restriction" group has been
enabled to enforce anonymous imports of traits imported only for their
methods (without being referenced explicitly) to make such import
statements more explicit - and update the code accordingly.

Lastly, this bumps `proptest-derive` from 0.4.0 to 0.5.0 because the old
version caused non-local `impl` errors on this new version of Rust.

--
1. https://blog.rust-lang.org/2024/11/28/Rust-1.83.0.html