Configure `--in-diff` to optimize mutation testing in CI
Update how `cargo mutants` is invoked in CI to optimize its use in
particular for Pull Requests, only testing what is relevant to the
diff.

This implementation maintains a full test run on pushes (to `main`)
by diffing against the hash for an empty commit, resulting in a diff
that represent the entire code base. The reason for doing a full
mutation test run on `main` is to ensure no mutants slip through[^1]
when the job duration does not matter.

^1: as described in <https://mutants.rs/in-diff.html>.