Bump Rust toolchain channel from 1.83 to 1.84
Upgrade the version of Rust and related tooling from 1.83.0 to 1.84.0,
which was recently released [1]. For Clippy, two new optional rules from
the "restriction" group has been enabled along with one moved there from
the "pedantic" group.

The lint `clippy::arbitrary_source_item_ordering` was not enabled since
I believe the enforced order is not beneficial to this project.

--
1. https://blog.rust-lang.org/2025/01/09/Rust-1.84.0.html