Bump Rust toolchain channel from 1.90 to 1.91
Upgrade the version of Rust and related tooling from 1.90.0 to 1.91.0,
which was recently released [1]. One new pedantic Clippy rule was added
which has been enabled.

The minimum `cargo-deny` dependency version was bumped from v0.18.0 to
v0.18.5 (latest) for compatibility, which also found one allowed license
is unused (hence it was removed).

-- 1. https://blog.rust-lang.org/2025/10/30/Rust-1.91.0/